refactor: 🚚 moved verification functions
