+++ src/app/api/admin/seed/route.ts
@@ -4,12 +4,12 @@ import { NextRequest, NextResponse } from 'next/server';
 
 import { connectDB } from '@/db/config';
 import Admin from '@/db/models/admin';
-import { verifyRequest } from '@/utils';
+import { verifyRequest } from '@/verify';
 
 export async function POST(request: NextRequest) {
   try {
     // Authenticates Request to see if it comes from DS3
-    const { isRequestValid, requestError } = verifyRequest(request);
+    const { isRequestValid, requestError } = verifyRequest(request.headers);
     if (!isRequestValid) return requestError;
 
     connectDB();
